refactor: moving cli print helpers to single internal package
